client/src/inference/requirements.txt
pybind11==2.13.1
numpy==1.21.4
tflite_runtime-2.3.1-cp37-cp37m-linux_armv6l.whl
